Factorise some common satellite functions
