+++ src/geefetch/data/satellites/gedi.py
@@ -254,19 +254,6 @@ class GEDIraster(SatelliteABC):
     def pixel_range(self):
         return 0, 100
 
-    def convert_image(self, im: ee.Image, dtype: DType) -> ee.Image:
-        match dtype:
-            case DType.Float32:
-                return im
-            case DType.UInt16:
-                min_p, max_p = self.pixel_range
-                return im.add(-min_p).multiply((2**16 - 1) / (max_p - min_p)).toUint16()
-            case DType.UInt8:
-                min_p, max_p = self.pixel_range
-                return im.add(-min_p).multiply((2**8 - 1) / (max_p - min_p)).toUint8()
-            case _:
-                raise ValueError(f"Unsupported {dtype=}.")
-
     def get_col(
         self, aoi: GeoBoundingBox, start_date: str, end_date: str
     ) -> ee.ImageCollection:
